+++ modules/v1_19_R3/src/main/java/net/countercraft/movecraft/compat/v1_19_R3/NextTickProvider.java
